requirements.txt
aiogram==3.26.0
aiohttp==3.10.11
aiosqlite==0.22.1
openpyxl==3.1.5
pillow==11.3.0
python-dotenv==1.1.1
